interface/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub==1.7.0
Cerberus==1.3.2
PyYAML==5.3.1
